Accept ImageStreamTags in ProfileBundles and use it
If an ImageStreamTag is given to a ProfileBundle's contentImage, all we
need to do is annotate the Deployment we create in order to
automatically track the ImageStreamTag and make use of the fancy
automatic updates it can do to deployments. This would ease the
operations for folks tracking their content this way.