Take into account transcient errors in profileparser
There could be temporary issues (e.g. ImagePullBackOff) that would
prevent the profileparser init container from running. This would make
the Deployment create a new pod in an attempt to fix the issue, and it
which could succeed. So, let's take into account only the newest pod
from the set of pods we fetch when validating the status. This way, the
deployment is able to recover.