Support disconnected installs
The current default behavior is to allow the `oscap` command to download
external resources. This is an issue in disconnected installs since
they'll either have no access to the internet, or have so, but through a
proxy.

This enables the usage of the operator in such environments by allowing
the admin to set two new scan settings:

* **noExternalResources**: This will entirely remove the flag that
  downloads external resources, and stuff like cve checks simply won't
  run.

* **httpsProxy**: This sets the `http_proxy` environment variable, so
  the `oscap` command will use the proxy. Note that the name
  `httpsProxy` was chosen to encourage folks to use https for their
  proxies.