Let RBAC issues slide when dealing with profilebundles
If the operator doesn't have permissions to access the bundle image,
this is not an issue. The operator will react as if the reference was
not an image stream and the error will ultimately manifest itself with
the container not being able to start... which will end up in the
profilebundle's status.