compliancescan: Don't keep checking the scan results if the CM is not ready
If the result ConfigMap was not created by the time we check the
aggregated results, we would previously just set the isReady flag, but
continue. The code would then proceed to the getScanResult function
which tries to check the CM's annotations and because the CM doesn't
even exist, getScanResult would have returned a failure. This means the
isReady flag would never be returned, but an error would be returned
instead.

This patch just shortcuts processing of the current configMap and moves on to
the next one. This way, if there is an error in the other CM's, it would have
bubbled up, but otherwise the isReady flag would have been returned as
expected.

Jira: OCPBUGSM-17900