compliancescan: Requeue if the result CM doesn't have the required annotations
We've seen cases where the results configMap exists, but does not have
the required annotations yet. Instead of returning a hard error in this
case, return the isReady flag set to false in gatherResults which would
cause the processing to be requeued again in a couple of seconds.

Jira: OCPBUGSM-17900