Hardening: Set some general hardening settings for our containers
This sets up some general hardening settings such as:

* Disallowing privilege escalation
* Read-only container filesystem

It also adds a read-only flag for a mount where needed.

This also mounts `/tmp` as an empty dir for the scanner containers. This
is because OpenSCAP will attempt to create a directory in `/tmp` in
order to temporarily dump the OVAL results. This way, we can mount the
directory and write to it, while still keeping the rest of the container
in a read-only filesystem.