Expose API resource collector errors as warnings
This exposes errors from the API resource collector by posting them
to a warnings file.

This will be read by the collector, uploaded to a ConfigMap, and
subsequently persisted to the Custom Resource.

The errors that we expose are:

* RBAC issues (Forbidden)
* Kind/Resource issues (if they don't exist)

This way we can indicate to users that they need to revisit their
permissions or install an appropriate operator.