e2e/rotation-checker: Change ImagePullPolicy to PullIfNotPresent
The testScheduledSuite tends to take a long time. We've noticed this
happens because it might take a long time to schedule the
rotation-checker pod. This pod will merely mount the PVC from the scan
and list the current results. We use this to verify that rotation has
indeed happened.

The hope is that by changing the rotation policy to PullIfNotPresent,
the image will already be in the node and the pod will have a faster
scheduling time.