Don't mark scan pods as still running when there's a failure
If there is a failure such as an eviction event, we don't want to mark
the pod as being done. Instead, let's mark it as if it's still running
and let the kubelet do its thing and restart the pod. Once the pod gets
rescheduled we should get the necessary results.

Co-Authored-By: Jakub Hrozek <jhrozek@redhat.com>