Only skip remediation application if it's a MachineConfig
We had a check in the aggregator that checked if a remediation can be
applied. This didn't take into account generic remediations, and so it
always skipped checks. It checked if we have set a specific nodeSelector
in the scan, and if so, it would apply the remediation. This removes this
requirement for objects other than Machineconfigs.