Add and fix relevant operator information in the CSV
The CSV can display relevant information about the operator. We weren't
using this feature... So let's do it!

This adds information about the operator itself and it's capabilities.
This will help users find the operator in the OperatorHub.

It also sets the "suggested-namespace" annotation, which enables the OLM
to automatically create the required namespace.

Also, it fixes the installation capabilities to reflect what the
operator actually supports.