Fix tailored Platform scans
We had code duplication between two functions that were setting up the
environment variables for the scans. When adding the tailoring support,
we forgot to add it to the function that sets up environment for the
platform scans.

This patch removes the code duplication and makes sure that the expected
environment is set up for both node and platform scans.