Funky fresh MCs
This uses the API reader client for exclusively for getting
MachineConfigs in the remediations controller. This is meant to reduce
the number of conflicts we get due to using a cached version of the
MachineConfig object we merge, since this always will query the API and
get the latest one.