Divide and Conquer (MachineConfigs)
The approach we used to have of attempting to do a big merge of
MachineConfigs is proving to be more trouble than its worth. The
rebels... I mean... MachineConfigs, are not gathered properly and
dissappear on reconcile loops.

Let's let Machine Config Operator do the merging, and only focus on
ensuring that the objects exist with the appropriate naming and labels.

This removes the extra MachineConfig logic in our reconciler, and
instead only handles validation and naming for MachineConfigs. Thus
simplifying the remediation controller.