Always update default profilebundles on operator startup
This ensures that the operator is using the ProfileBundle image that
it's supposed to be using. The intent of this patch is to help on
updates.

The workflow is as follows:

* An initial PB is created with a certain tag
* When an update happens, the Deployment rolls out and restarts the
operator
* The operator will check the contentImage reference and update (by
  patching it) if needed.

This way we ensure that folks are using supported images and not modify
the default profile bundles.

Note that with this change, any error when creating the ProfileBundles
will result in the operator exiting and retrying.