refactor: Introduce scan type handler
This splits the scan handling in a per-type object. This way we won't
have the big switch code blocks that we used to, and this will be done
in object-specific functions.

The node scan type handler will gather the relevant nodes on the
validation phase (which now happens before any phase in the reconcile
loop) and so it'll always have that reference of those objects when it
needs it.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>