Add annotations to better deal with remediations
This adds a couple of remediations that ComplianceSuites react to in
order to better deal with ComplianceRemediations:

* `compliance.openshift.io/apply-remediations`: Applies all of the
remediations created by a ComplianceSuite's Scans. It's similar to the
`AutoApplyRemediations` flag, but it's meant to be a one-shot operation,
which removes the annotation once it's done applying.

* `compliance.openshift.io/remove-outdated`: Can be used if either the
`apply-remediations` annotation is used, or the `autoApplyRemediations`
flag is enabled. It will automatically remove the outdated remediations
objects in order to update them (leaving the current ones to be applied).

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>