Use a separate SA for resultserver
resultserver was relying on the "default" SA to apply the restricted SCC on
the resultserver pod. While this is OK in the general case, in environments
that bound the "default" SA to some other SCC than restricuted, this would
have broken the assumption by applying a different SCC and not propagating
settings like fsGroup to the pod.

Jira: OCPBUGSM-23632