Replace xmldom by xmlquery for profile parsing
xmldom seems to be a stale project, and its current issues present some
usability limitations for us:

* Excessive resource usage
* Inability to get full xml content from sub-nodes.

The second point results in us getting malformed descriptions in some
rules.

This aims to address that by replacing the XML parsing for another
library that's more active.

This is only used when generated profiles, rules and variables.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>