Add warnings to ComplianceCheckResult
Lately, we've gotten reports of folks getting confused by some results
being NOT-APPLICABLE for some benchmarks.

Even if for us this is normal and expected, users find it confusing and
will think something is wrong.

In a recent PR we also are handling rules differently depending on the
OpenShift version [1]. So, this makes it very relevant for us not to
confuse users and try to give as much relevant information as possible.

To address this, the version dependency notice has been added as a
warning in the rule itself. To reflect this in the operator, this
commit adds a `warnings` section to the ComplianceCheckResult Custom
Resource. This goes through the warnings and exposes them to users,
hence explaning possible things people want to watch out for.

Note that we are using the warnings for persisting the API path
information. This is not very relevant to users, so this is skipped when
adding it to the result.

I also noticed at we were only taking one warning into account when
parsing the aforementioned API path; this fixes that. Only waking one
warning into account could have been problematic in cases where the API
path is not present in the first warning of a rule.

[1] https://github.com/ComplianceAsCode/content/pull/6813

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>