Converge rules that present multiple results
This enables the Compliance Operator to handle rules that output
multiple results.

This is the case for rules like `security_patches_up_to_date` which will
present a result per-package.

For this kind of results, we get one final result. So logic to get the
most relevant result was added. Being `ERROR` the most relevant result,
followed by `FAIL`.

To facilitate this, the results are now stored in memory in a hash, as
opposed to the previous list. This allowed for us to remove the Id from
the structure of ParsedResults and leverage the id that's stored in the
hash already.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>