Add `failureInfo` to handle extra information for rules
This new parameter in the ComplianceCheck result will be used to gather
failure information that could be useful to asses the results given by
the operator.

Right now, its only usage is for verifying the
`security_patches_up_to_date` rule, which is a `multi-check` of OVAL
class "patch".

As we get requests, we'll add handling for different class types.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>