Make use of the HTTPS_PROXY environment variable
OpenShift allows the administrator to set a global proxy by editing the
proxy/cluster object:
    https://github.com/openshift/api/blob/master/config/v1/types_proxy.go
The OLM can in turn watch for the proxy objects and set the HTTPS_PROXY
environment variable in the CO deployment.

This patch sets the proxy environment variable for the openscap pods
based on the value of the HTTPS_PROXY environment variable passed for
the operator, thus taking advantage of the global proxy settings.

The HTTPSProxy member of the ComplianceScanSettings structure is
retained, just an attribute is added that the OCP proxy is preferred.

Jira: CMP-846