Limit resource usage for all workloads
Now that our profiles aren't using as much resources as they used to,
let's add appropriate requests and limits for all resources.

The max CPU usage that a workload is allowed to have for the compliance
operator is 100m, this is with the intent of leaving resources for other
workloads so this would work in single node deployments.

This minifies the CPU usage for ALL workloads.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>