Die xmldom! Die!
This PR removes go-xmldom as a dependency, finally switching everything
to use xmlquery instead. This gets rid of some parsing issues that
prevented users from seeing the whole text of instructions, descriptions
and other relevant bits of our content in the generated Kubernetes
objects.

In order to make this happen, we now had to take into account the XML
namespaces in objects.

For testing, I had to update a datastream file we were using... which
really wasn't a datastream, but a result. ARF generated by openscap
doesn't preserve the namespaces. So I updated the testing XML which
gives us more relevant results anyway.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>