Convert XML from CaC content to markdown in the k8s objects
Some elements contain HTML tags in the content and we used to display
them verbatim, which was hard to read. This patch uses the
github.com/jaytaylor/html2text library to convert HTML to markdown and
display that instead.

This patch renders:
    Rule:
        - rationale
        - description
        - warning
    Profile:
        - description
    Variable:
        - description

Variable values were left out on purpose; in theory they could contain
XML and in general we should treat the variable values as opaque.

Note that the OCIL elements that we put into ComplianceCheckResult's
instructions attribute were not converted. Despite CaC using html tags
in the OCIL attribute, the DataStream already converts OCIL into plain
text.

Because the parsing is now a little slower, I've bumped the deadline
for parsing the profiles from 1.0 to 1.2

Jira: CMP-925