Collect all ocp-api-endpoint elements
Some rules need to collect several resources and correlate them or run a
check on several results. For example OAuth token inactivity can be set
either in the OAuth server object or any of the OAuth clients.

This patch extends gathering the API resource URLs to all of those that
are included in the warnings elemement instead of the first one to
support cases like above.

The functionality to add multiple URLs to a single warning is included
with https://github.com/ComplianceAsCode/content/pull/7089

Jira: CMP-775