Restart profileparser on failures
We used to be too permisive about failures in the profileparser, which
worked fine in the beginning, but doesn't work so well since we started
annotating the objects with the bundle SHA digest during an update and delete
any objects that were not annotated with that digest value - in case
an update operation failed because of a transient error, we would have
skipped annotating that object and ultimately treat it as removed and
delete it. This was happening especially during upgrades where the API
server might be unreachable at some point.

This patch exits the profileparser with a non-zero code on failures as
well as propagates the errors from the client operations to the error
channels so that they are eventually returned to the binary and used for
the error code. This ensures that the profileparser restarts and
hopefully goes through after the restart.

Jira: OCPBUGSM-32701