Enable Creation of TailoredProfiles without extending existing ones
This removes the requirement of having to extend a Profile in order to
create a TailoredProfile. So, the `extends` field from the
TailoredProfile CRD is no longer mandatory.

This requires for users to only select rules from the same
ProfileBundle, as the ProfileBundle will become "owner" of the
TailoredProfile object (for garbage collection reasons).

Also, TailoredProfiles now require the following annotation (which is
currently used by Profiles):

    compliance.openshift.io/product-type: <Type>

If left unset, the default will be `Platform`.

If the TailoredProfile's name ends with the suffix `-node`, the
annotation will have the value `Node`, and will schedule a `Node` scan
type instead.

If the annotation is set, it's left untouched.

This is a little manual right now, but can be extended in the future.

The ScanSettigBinding controller will also now wait for a
TailoredProfile object to be parsed and ready before trying to create a
ComplianceSuite.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>