e2e: Migrate TestScanProducesRemediations to use ScanSettingBinding
This migrates the test to use ScanSettingBindings and TailoredProfiles,
which is what we recommend users do. It also stops using the
`rhcos4-moderate` profile which is very time and resource intensive, and
instead uses the `ocp4-moderate` which is quite complete but lighter.

The profile change also fixes a fluke. The moderate profile doesn't
produce all remediations with the "applicable" state, as some need
dependencies. So we couldn't even do this test properly with that
profile anyway. So the profile change is needed.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>