aggregator: Remove MachineConfig validation
we don't need to do this in the aggregator, the validation is ran anyway
on the remediation controller. This way we can actually show to the user
why a MachineConfig remediation wasn't able to be created.

With this, the aggregator is just a "simple" program that aggregated and
creates remediations, no extra magic.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>