Validate that rules in tailored profile are of appropriate type
This PR now validates that the rules added to a tailored profile belong
to a relevant type.

As we know, we have the following types:

* **Platform**: Runs on the platform, do k8s checks.
* **Node**: Runs on the node itself.
* **None**: Informational.

When building a tailored profile, we want there only to be rules that
are gonna be relevant and run as the user expects. Here, you can mix and
match the following:

* Platform rules and None rules
* Node rules and None rules

but nothing in between. Platform and Node rules shouldn't mix.

With this validation, users will now see an error in the tailored
profile and will not build tailored profiles that don't do what they
want.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>