Revamp CRD docs
This adds pictures and changes the ordering to show how we expect users
to interact with the operator. This way, suites and scans are no longer
on the top, and instead the more relevant objects to interact with are
there.

It also divices the document in relevant sections depending on the
classification of the object.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>