Result Server: Make nodeSelector and tolerations configurable
This makes the aforementioned parameters configurable through the
ScanSettings object.

This enables deployers to configure where the Result Server will run,
and thus what node will host the Persistent Volume that will contain the
raw results.

This is needed for cases where the storage driver doesn't allow us to
schedule a pod that makes use of a persistent volume on the master
nodes.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>