Add option to make scan scheduling strict/not strict
Currently, when dealing with node scans, we only allow the scan to go
forward if the node is not unschedulable (via the unschedulable flag of
the Node object). However, for public cloud cases where there might be a
lot of ephemeral nodes (e.g. when a node autoscaler is being used), this
is not ideal as it's very difficult to do scans. On the other hand, if a
node became unready while the scan was on-going, we also return a
similar error.

So, to deal with this, the PR introduces the option `strictNodeScan`,
which enables an administrator to toggle whether we want to be strict
when getting results of a scan.

The default is `true`, which is the setting we were using before.
However, when set to `false`, the scan will not fail if a node is set to
`unschedulable` and will ignore "unschedulable" errors on the scan pods.
Thus allowing us to still get a subset of scans.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>