Implement version applicability for remediations
This PR enables us to specify the OCP and Kubernetes version
applicability through annotations in the remediation objects.

This will read the following annotation keys:

    complianceascode.io/ocp-version

and

    complianceascode.io/k8s-version

Which allow us for defining ranges in semver format. e.g. `>4.10.1`

This is done through the `github.com/blang/semver/` library, which has
further descriptions about the range format.

When detecting the aforementioned annotations, they'll be transformed to
Compliance Operator equivalent ones. e.g. using the
`compliance.openshift.io/` prefix, as opposed to `complianceascode.io/`.

When a remediation is skipped due to this mechanism, an appropriate
event will be generated.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>