Add a more verbose Changelog for the recent versions
Our changelog is just a glorified git log. This is useful to get an
overview of the changes for people who are connected to the development,
but not for anyone outside.

This patch makes the changelog more verbose for the last couple of
versions, since the last one that OpenShift had shipped.