Inherit scheduling for workloads from Compliance Operator manager
When the Compliance Operator was first coded, some assumptions were made
that were specific to Openshift. One of these assumptions was the fact
that the `node-role.kubernetes.io/` was more actively used amongst other
distros. An even worst assumption was that the `master` role was always
available. These statements hold true for OpenShift, but not for other
distros like EKS.

This PR removes that assumption by making every workload inherit the
same `nodeSelector` and `tolerations` from the controller manager. This
is done by passing this information to every controller. However, using
the information is optional.

When CO begins, the manager will query it's own pod and fetch this
information. Failure to do so will result in an error.

Since we fetch the operator name and namespace from constants, the
constants were modified to not depend any more on k8sutils. the
aforementioned library will no longer be available in further versions
of operator-sdk, so it's better to move away from it now little by
little.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>