Introduce process documentation for maintaining release notes
We have some automation in place that helps us generate release notes.
While tooling is important for maintaining a project, we still need a
human to verify the release notes make sense. Additionally, we don't
have specific guidelines that describe what we expect for a release
note, or guide contributors on how to write them.

This commit attempts to add documentation to the contributor guide that
introduces those conventions, provides examples for what we expect in
our CHANGELOG, and describes the overall process for maintaining it.