Refactor update_changelog.sh to follow release note process
Recent changes to our release note process requires reviewers to review
release notes while the change they're describing is under review. This
shifts the process of writing the release note left, instead of a
retroactive activity the maintainers need to do after the change merged.

This script updates the update_changelog.sh script to properly label the
unreleased notes with a release version and also generates a new set of
release note templates to use for the next, upcoming release.