Make sure the helm chart.yaml changes are included in release commit
Since the release tooling will update the `appVersion` in
deploy/compliance-operator-chart/Chart.yaml, we should make sure we add
that file to the release commit.

This commit also simplifies the sed regular expression for updating the
appVersion in deploy/compliance-operator-chart/Chart.yaml so that it
doesn't try to match the end of the line. This was causing the bundle
make target to fail because it escaped the forward slash denoting the
other part of the sed command. Here, let's just simplify the expression
to not include $ since it's not necessary anyway.