Break the release process apart
The `make release` target lumped the entire release process together,
which was nice for simplicity, but it doesn't give you, the person
proposing the release, an opportunity to review the changes before
they're proposed.

In the event something is incorrect (e.g., if the operator-sdk did
something we didn't expect), we should have the ability to catch that
and fix it before proposing the release. Otherwise, we will need to fix
upstream branches and update outdated tags.

Fixes #783