Remove the deploy/eks.yaml deployment
This deployment was added recently so that we could deploy the
compliance-operator on EKS. Shortly after, we start implementing a Helm
chart for deploying the operator on both OpenShift and EKS, depending on
the values passed into Helm.

Removing this file since it's existence is redundant with what Helm
provides and it breaks `make deploy-local` as well as the OLM artifacts
during the release process.

Fixes #781