e2e: Don't rely on order of inconsistent results
The TestInconsistentResult e2e test was relying on a certain order in
which a test result is labeled. If the result order was swapped, even if
it contained the same results, the test would fail.

This patch fixes the test to accept either of the two possible
orderings. Since there's only two, both are hardcoded and checked, if we
ever have more combinations, we'll refactor the code so that it is more
generic.