scans: Delete scan pods and aggregator when they're done
We didn't delete scan pods and the aggregator after a scan was done
unless the delete was forced by the scan object being deleted. This
appears to be breaking the cluster-autoscaler which can't drain pods
from nodes that are not backed by a replicating controller.

Since there appears to be no reason to keep the pods around unless the
debug=true option is set, let's remove those pods when the scan reaches
done and only keep the other resources around that are needed for
eventually re-running the scan.

Jira: OCPBUGSM-43262