api-resource-collector: Use paging and filter file contents when listing MachineConfigs
We have seen reports of the api-resource-collector being killed when
used for the PCI-DSS standard. Turns out that the PCI-DSS content lists
all MachineConfigs and on very large clusters with many MachineConfigs,
this can amount to tens of megabytes that the api-resource-collector
would then keep in memory.

Because we don't intend to inspect file contents of MCs (this is what
the OS-level checks are for), let's strip the file contents out of MCs.
In order to keep everything in memory before the files are filtered out,
this patch also uses paging using the "limit" and "continue" query
parameters of k8s API as described here:
    https://kubernetes.io/docs/reference/kubernetes-api/common-definitions/list-meta/

Jira: OCPBUGSM-45245