changelog: Point out important fixes to OCP4/RHCOS4 compliance content
Historically, we've only been including CO fixes in CO's changelog. This
was, while technically correct, confusing to users, who tend to consume
the operator and the content as a single unit.

Let's mention the important content fixes and improvements in a
sub-section of the CHANGELOG.md file to better inform our users about
what's new.