Ignore tests/_setup directory
The operator-sdk update refactored how we setup end-to-end testing. Part
of that was to include a separate directory for custom resource
defintions and custom resources generated by the tooling.

Since the files in tests/_setup are generated, we don't need to track
them in the project. Let's add them to .gitignore.