Make scanner's and api-resource-collector's limits configurable
Even with the recent fix to the memory consumption, we might need to
provide a way for our users to bump the limits on scans, especially with
huge cluster nodes and Linux scans that traverse the filesystem.

This patch adds a new scanLimits attribute to the scan settings CR.
If unset, the compiled-in defaults are used, otherwise the settings take
precedence.

The only containers that are affected are the scanner and the
api-resource-collector as we've seen reports with those containers being
OOMKilled.

Jira: OCPBUGSM-39585