e2e: Remove node taint in a closure
We had issues where a taint was not removed correctly. This lead to
cascading failures in the e2e suite where all subsequent test would
fail.

Even if removing the taint in a defer might not help prevent the
original failure, it would help prevent the cascade and make it easier
to detect issues in our e2e suite.