Allow running a subset of tests through a MakeFile variable
This introduces the `E2E_TEST_TYPE` variable, which, if set, allows the
e2e test suite to run only a subset of tests. This subsets are defined
in-code and currently only take either `parallel` or `serial`. However,
the default is to run them all (the default is literally called `all`)
so the current behavior is kept.

This will allow us to create another test in the `openshift/release`
repo and divide the test runs to take different clusters, which will
speed up the e2e CI run.

We could have `e2e-serial` and `e2e-parallel` tests.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>