scan: Add conditions to scans
Some users were confused about our re-running feature. They expected the
suites to be re-created every time and the fact that they were not and
the createTimestamp remained constant was causing questions.

Let's add conditions to the scans as well so that it's obvious that the
scan was re-ran.

Jira: OCPBUGS-379