Reproducible builds: Ensure we can get reproducible builds for CO
By setting the `-trimpath` flag, we ensure that we don't pollute the
debug paths with absolute paths of the environment where we built the
image; instead, this will will show the path of the go module (which is
also quite understandable for debugging). Making the go binary's buildid
blank also ensures that the binary has data we can predict, and thus
ensures reproducible builds.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>