Resultserver rotation test timing
The last resultserver unit test (rotation of 3 directories with
policy=2) often failed on ppc64le systems. It was observed that test
directories created only within 5 milliseconds of each other could
result in an identical creation time, and affect the expected results.

This increases the delay between test directories to 100ms to better
ensure a different creation time on ppc64le.