Add a script to conditionally set upstream remote
Our upstream release process expects maintainers to propose release
patches, branches, and tags to the upstream repository
(ComplianceAsCode/compliance-operator). Some people prefer to setup
their upstream remote as origin, while others use upstream.

This change introduces a script to determine which git remote to use for
the GIT_REMOTE variable based on the url associated with the upstream
repository. This makes the release process consistent, regardless of who
is doing the release. This commit also removes an extra GIT_REMOTE
defintion in the Makefile, making it clear where we're setting that
variable and what the intended value should be.